@@ -1,13 +1,13 @@
 name: flutter_playout
 description: This plugin provides audio/video playback from a URL with background audio support, lock screen
   controls & player events for both iOS and Android.
-version: 1.0.40
-author: Muslim Television Ahmadiyya International <developer@mta.tv>
+version: 2.0.1
 homepage: https://www.mta.tv
 repository: https://github.com/muslimtv/flutter_playout
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=1.12.13+hotfix.6"
 
 dependencies:
   flutter:
@@ -27,8 +27,12 @@ flutter:
   # be modified. They are used by the tooling to maintain consistency when
   # adding or updating assets for this project.
   plugin:
-    androidPackage: tv.mta.flutter_playout
-    pluginClass: FlutterPlayoutPlugin
+    platforms:
+      android:
+        package: tv.mta.flutter_playout
+        pluginClass: FlutterPlayoutPlugin
+      ios:
+        pluginClass: FlutterPlayoutPlugin
 
   # To add assets to your plugin package, add an assets section, like this:
   # assets:
