@@ -6,7 +6,7 @@ import 'package:flutter/services.dart';
 /// for HLS manifest. Use [setPreferredAudioLanguage] to set/change
 /// language for the currently playing asset.
 mixin MultiAudioSupport {
-  MethodChannel _methodChannel;
+  MethodChannel? _methodChannel;
   Future<void> enableMultiAudioSupport(int viewId) async {
     _methodChannel =
         MethodChannel("tv.mta/NativeVideoPlayerMethodChannel_$viewId");
@@ -16,11 +16,8 @@ mixin MultiAudioSupport {
   /// is based on ISO_639_2 language codes. Please see
   /// [lib/multiaudio/ISO_639_2_LanguageCode.dart]
   void setPreferredAudioLanguage(String languageCode) async {
-    if (_methodChannel != null &&
-        languageCode != null &&
-        languageCode.isNotEmpty &&
-        !Platform.isIOS) {
-      _methodChannel.invokeListMethod(
+    if (_methodChannel != null && languageCode.isNotEmpty && !Platform.isIOS) {
+      _methodChannel!.invokeListMethod(
           "setPreferredAudioLanguage", {"code": languageCode});
     }
   }
