@@ -1,5 +1,3 @@
-import 'package:flutter/cupertino.dart';
-
 /// Used for adding text tracks to the player. [mimetype] is the type of text
 /// track for example text/webvtt. [languageCode] is the language of the track
 /// for example `en` or `fr` and [uri] is the location for the track.
@@ -11,9 +9,9 @@ class TextTrack {
   TextTrack(this.mimetype, this.languageCode, this.uri);
 
   factory TextTrack.from(
-      {@required String mimetype,
-      @required String languageCode,
-      @required String uri}) {
+      {required String mimetype,
+      required String languageCode,
+      required String uri}) {
     return new TextTrack(mimetype, languageCode, uri);
   }
 
