@@ -28,24 +28,24 @@ mixin PlayerObserver {
 
   /// Override this method to get update when playhead moves. This method
   /// fires every second with [position] as seconds.
-  void onTime(int position) {/* user implementation */}
+  void onTime(int? position) {/* user implementation */}
 
   /// Override this method to get notifications when a seek operation has
   /// finished. This will occur when user finishes scrubbing media.
   /// [position] is position in seconds before seek started.
   /// [offset] is seconds after seek processed.
-  void onSeek(int position, double offset) {/* user implementation */}
+  void onSeek(int? position, double offset) {/* user implementation */}
 
   /// Override this method to get notifications when media duration is
   /// set or changed.
   /// [duration] is in milliseconds. Returns -1 for live stream
-  void onDuration(int duration) {/* user implementation */}
+  void onDuration(int? duration) {/* user implementation */}
 
   /// Override this method to get errors thrown by the player
-  void onError(String error) {/* user implementation */}
+  void onError(String? error) {/* user implementation */}
 
   void _processEvent(dynamic event) async {
-    String eventName = event["name"];
+    String? eventName = event["name"];
 
     switch (eventName) {
 
@@ -73,7 +73,7 @@ mixin PlayerObserver {
       case "onSeek":
 
         /* position of the player before the player seeks (in seconds) */
-        int position = (event["position"]).toInt();
+        int? position = (event["position"]).toInt();
 
         /* requested position to seek to (in seconds) */
         double offset = double.parse("${event["offset"]}");
